refactor(lap-detector): named detector IDs + per-detector versioning
- LAP_DETECTOR_ID = "lapdetector_v1" (FM2023, F1 — LapDetector)
- LAP_DETECTOR_V2_ID = "ac_lapdetector_v2" (ACC, AC Evo — LapDetectorV2)
- Rename lap-detector-v2.ts → lap-detector-ac.ts
- Add detectorId to ILapDetector interface; each class implements it
- Pipeline uses detector.detectorId instead of imported constant
- reprocessSession uses detector.detectorId after replay
- countStaleSessions/getStaleSessions accept string[] of valid IDs
- Startup and routes pass both IDs for stale detection
