+++ server/index.ts
@@ -199,13 +199,15 @@ udpListener.start(udpPort);
 // Stores the notification in wsManager so it's sent to each client on connect.
 import { countStaleSessions } from "./db/queries";
 import { LAP_DETECTOR_ID } from "./lap-detector";
-countStaleSessions(LAP_DETECTOR_ID).then((count) => {
+import { LAP_DETECTOR_V2_ID } from "./lap-detector-ac";
+const ALL_DETECTOR_IDS = [LAP_DETECTOR_ID, LAP_DETECTOR_V2_ID];
+countStaleSessions(ALL_DETECTOR_IDS).then((count) => {
   if (count > 0) {
     console.log(`[Server] ${count} session(s) recorded with stale lap detector — will prompt user to reprocess`);
     wsManager.setStaleSessionsNotification({
       type: "stale-lap-detection",
       sessionCount: count,
-      currentVersion: LAP_DETECTOR_ID,
+      currentVersion: `${LAP_DETECTOR_ID},${LAP_DETECTOR_V2_ID}`,
     });
   }
 }).catch((err) => {
